requirements.txt
anyascii==0.1.7
appdirs==1.4.4
arabic-reshaper==2.1.0
asgiref==3.4.1
astroid==2.3.3
autopep8==1.5.7
beautifulsoup4==4.8.2
birthdays==0.4.4
bootstrap3-datetime==2.4
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
class-registry==2.1.2
click==7.1.2
colorama==0.4.4
cryptography==3.2
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.13
django-admin-autocomplete-filter==0.6.1
django-allauth==0.42.0
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-autoslug-iplweb==1.9.4.dev0
django-birthday==0.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap3==14.0.0
django-bootstrap3-datetimepicker-2==2.8.2
django-bootstrap4==2.1.1
django-braces==1.14.0
django-cities-light==3.7.0
django-classy-tags==2.0.0
django-clear-cache==0.3
django-cms==3.8.0
django-cors-headers==3.6.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extra-views==0.13.0
django-filter==2.4.0
django-formset-bootstrap==1.4.0.dev3
django-formtools==2.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-import-export==2.4.0
django-jquery==3.1.0
django-localflavor==3.0.1
django-markdown2==0.3.1
django-model-utils==4.0.0
django-modelcluster==5.1
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
django-sb-admin==5.5.1.1
django-sekizai==2.0.0
django-smart-selects==1.5.9
django-sslcommerz==1.0.0a0
django-static-jquery==2.1.4
django-tables2==2.3.1
django-taggit==1.3.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangocms-admin-style==2.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
filelock==3.0.12
filters==1.3.2
filters-django==1.0.5
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.8
importlib==1.0.4
isort==4.3.21
jdcal==1.4.1
jedi==0.17.2
Js2Py==0.70
jsonify==0.5
l18n==2020.6.1
lazy-object-proxy==1.4.3
markdown2==2.3.9
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.4
parso==0.7.1
pilkit==2.0
Pillow==8.1.0
pipenv==2022.1.8
pipwin==0.5.0
portalocker==2.0.0
progressbar2==3.52.1
psycopg2 
pycodestyle==2.7.0
pycparser==2.20
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pypandoc==1.5
pyparsing==2.4.7
PyPDF2==1.26.0
PyPrind==2.11.2
pysaml2==6.2.0
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python-stdnum==1.13
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
pywin32==300
PyYAML==5.3.1
regex==2020.10.23
reportlab==3.5.54
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.2
typed-ast==1.4.2
tzdata==2022.1
tzlocal==2.1
Unidecode==1.1.1
Unipath==1.1
urllib3==1.25.7
virtualenv==20.1.0
virtualenv-clone==0.5.7
wagtail==2.11.2
wagtailfontawesome==1.2.1
waitress==1.4.4
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
wrapt==1.11.2
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
zope.interface==5.1.2
